Updated: directory restructuring
