+++ native/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
-import OnboardingNavigator from "./src/feature/navigation/navigator/OnboardingNavigator";
+import OnboardingNavigator from "./src/navigator/OnboardingNavigator";
 
 export const App: React.FC<{}> = () => {
   return (
